==> 20_preprocess.ipynb <==
from pathlib import Path
import os
import sys
import gc
import re
import shutil
import json
import math
import jinja2
from collections import defaultdict
import numpy as np
import pandas as pd
import bitsandbytes
import accelerate
import torch
from transformers import AutoTokenizer, AutoModelForSequenceClassification
from tqdm import tqdm
import matplotlib.pyplot as plt
from typing import Dict, List, Tuple, NamedTuple, Callable, Iterable, Set, Optional, Any
import scml
from scml import pandasx as pdx
import lalaes2 as mylib
print(f"accelerate={accelerate.__version__}, bitsandbytes={bitsandbytes.__version__}")

version = "02"
corpus_key = "comp"
corpus_map = {
    "comp": Path("input/train.csv"),
    "persuade": Path("input/persuade20/persuade_2.0_human_scores_demo_id_github.csv"),
}
# model for pseudo labelling Persuade 2.0 topic
directory = Path("models/persuade/deberta_v3_base/20240626_132806")
model_max_length = 512
basic = mylib.BasicPreprocessor()
bow = mylib.BowPreprocessor(drop_stopword=True)

tim = scml.Timer()
tim.start()
percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
os.environ["TOKENIZERS_PARALLELISM"] = "false"
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()
scml.seed_everything()

device = torch.device("cpu")
if torch.cuda.is_available():
    device = torch.device("cuda:1")
    for i in range(torch.cuda.device_count()):
        print(f"device={i}, {torch.cuda.get_device_name(i)}")
        print('Mem Allocated:', round(torch.cuda.memory_allocated(i)/1024**3,1), 'GB')
        print('Mem Cached:   ', round(torch.cuda.memory_reserved(i)/1024**3,1), 'GB')
else:
    print("cpu")

df = pd.read_csv(corpus_map[corpus_key], low_memory=False)
df.info()

if corpus_key=="persuade":
    df = df.rename(columns={
        "essay_id_comp": "essay_id",
        "holistic_essay_score": "score",
        "prompt_name": "topic",
        "assignment": "prompt",
    })
cols = ["score"]
df[cols] = df[cols].astype(np.int8)

def preprocess_text(fn, col) -> Callable:
    def inner(row) -> str:
        return fn(row[col])
    
    return inner


cols = ["full_text"]
for col in cols:
    print(col)
    df[col] = df.progress_apply(preprocess_text(basic, col), axis=1)

# Pseudo Labelling Topic

Classify text to 1 of 15 topics from Persuade 2.0 corpus.

if corpus_key=="comp":
    tokenizer = AutoTokenizer.from_pretrained(directory, model_max_length=model_max_length)
    print(f"""{tokenizer}
    model_input_names={tokenizer.model_input_names}
    pad_token_id={tokenizer.pad_token_id}
    """)

if corpus_key=="comp":
    model = AutoModelForSequenceClassification.from_pretrained(directory)
    print(model)
    print(model.config)

if corpus_key=="comp":
    y_pred = np.argmax(
        mylib.classify_persuade_topic_logits(
            ds=mylib.PersuadeDataset(
                tokenizer=tokenizer,
                texts=df["full_text"].tolist(),
            ),
            model=model,
            batch_size=128,
            device=device,
            progress_bar=True,
        ),
        axis=1,
    ).flatten()
    df["topic"] = [mylib.PersuadeDataset.ID_TO_LABEL[i] for i in y_pred]

cols = ["topic"]
for col in cols:
    print(col)
    df[col] = df.progress_apply(preprocess_text(bow, col), axis=1)

pdx.value_counts(df["topic"])

# Review Data

cols = ["essay_id", "score", "topic", "full_text"]
df = df[cols]
df.info()

df.head()

df.to_parquet(f"output/{corpus_key}_{version}.parquet", index=False)
assert df.notna().all(axis=None)

tim.stop()
print(f"Total time taken {str(tim.elapsed)}")
