==> 21_llm_critique.ipynb <==
from pathlib import Path
import os
import sys
import gc
import re
import shutil
import json
import math
import jinja2
from collections import defaultdict
import numpy as np
import pandas as pd
import bitsandbytes
import accelerate
import torch
from transformers import AutoTokenizer, AutoModelForCausalLM, BitsAndBytesConfig
from tqdm import tqdm
import matplotlib.pyplot as plt
from typing import Dict, List, Tuple, NamedTuple, Callable, Iterable, Set, Optional, Any
import scml
from scml import pandasx as pdx
import lalaes2 as mylib
print(f"accelerate={accelerate.__version__}, bitsandbytes={bitsandbytes.__version__}")

class ModelConf(NamedTuple):
    directory: Path
    model_max_length: int
    char_limit: int
    load_in_dtype: str = "auto"


#input_filename = "persuade_02.parquet"
input_filename = "comp_02.parquet"
template_version = 3
llm_key = "Qwen2-1.5B-Instruct"
llm_map: Dict[str, ModelConf] = {
    "gemma-1.1-2b-it": ModelConf(
        directory=Path("huggingface/google/gemma-1.1-2b-it"),
        model_max_length=8192,
        char_limit=3000,
        load_in_dtype="fp16",
    ),
    "Qwen2-1.5B-Instruct": ModelConf(
        directory=Path("huggingface/Qwen/Qwen2-1.5B-Instruct"),
        model_max_length=8192,
        char_limit=3000,
    ),
}
mc = llm_map[llm_key]
critique_col = f"ctq_{template_version}_{llm_key}"
environment = jinja2.Environment()
basic = mylib.BasicPreprocessor()

v1_template = environment.from_string(
"""You are a teacher grading a student's essay. Assign a score on a scale of 1 to 6 based on the following rubric. A score of 6 indicates clear and consistent mastery with minor errors, showcasing insightful development of a viewpoint, outstanding critical thinking, strong use of appropriate examples and evidence, well-organized structure with coherence and smooth progression of ideas, skillful language use with varied vocabulary and sentence structure, and minimal errors in grammar, usage, and mechanics. Scores decrease with occasional errors (score of 5), lapses in quality (score of 4), weaknesses in critical thinking and organization (score of 3), serious flaws in viewpoint and coherence (score of 2), and fundamental errors hindering meaning (score of 1).
In your response, output only the score without explanation.
[ESSAY] {{ essay }} [/ESSAY]
"""
)
v2_template = environment.from_string(
"""You are a teacher grading a student's essay. Assign a score on a scale of 1 to 6 based on the following rubric. A score of 6 indicates clear and consistent mastery with minor errors, showcasing insightful development of a viewpoint, outstanding critical thinking, strong use of appropriate examples and evidence, well-organized structure with coherence and smooth progression of ideas, skillful language use with varied vocabulary and sentence structure, and minimal errors in grammar, usage, and mechanics. Scores decrease with occasional errors (score of 5), lapses in quality (score of 4), weaknesses in critical thinking and organization (score of 3), serious flaws in viewpoint and coherence (score of 2), and fundamental errors hindering meaning (score of 1).
Provide constructive feedback to improve the essay.
[ESSAY] {{ essay }} [/ESSAY]
"""
)
v3_template = environment.from_string(
"""You are a teacher grading a student's essay. Assign a score on a scale of 1 to 6 based on the following rubric. A score of 6 indicates clear and consistent mastery with minor errors, showcasing insightful development of a viewpoint, outstanding critical thinking, strong use of appropriate examples and evidence, well-organized structure with coherence and smooth progression of ideas, skillful language use with varied vocabulary and sentence structure, and minimal errors in grammar, usage, and mechanics. Scores decrease with occasional errors (score of 5), lapses in quality (score of 4), weaknesses in critical thinking and organization (score of 3), serious flaws in viewpoint and coherence (score of 2), and fundamental errors hindering meaning (score of 1).
In your response, state the score and areas for improvement.
[ESSAY] {{ essay }} [/ESSAY]
"""
)
template = v1_template
max_new_tokens = 16
if template_version==2:
    template = v2_template
    max_new_tokens = 64
if template_version==3:
    template = v3_template
    max_new_tokens = 64

tim = scml.Timer()
tim.start()
percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
os.environ["TOKENIZERS_PARALLELISM"] = "false"
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()
scml.seed_everything()

device = torch.device("cpu")
if torch.cuda.is_available():
    device = torch.device("cuda:1")
    for i in range(torch.cuda.device_count()):
        print(f"device={i}, {torch.cuda.get_device_name(i)}")
        print('Mem Allocated:', round(torch.cuda.memory_allocated(i)/1024**3,1), 'GB')
        print('Mem Cached:   ', round(torch.cuda.memory_reserved(i)/1024**3,1), 'GB')
else:
    print("cpu")

df = pd.read_parquet(Path(f"input/{input_filename}"))
#df = df.sample(frac=0.001)
df.info()

# LLM-as-a-Judge

LLM takes on the role of a teacher to generate essay critique.

tokenizer = AutoTokenizer.from_pretrained(mc.directory, model_max_length=mc.model_max_length)
tokenizer.pad_token = tokenizer.eos_token
print(f"""{tokenizer}
model_input_names={tokenizer.model_input_names}
pad_token_id={tokenizer.pad_token_id}
""")

%%time
print(f"Load in {mc.load_in_dtype}")
if mc.load_in_dtype=="fp16":
    model = AutoModelForCausalLM.from_pretrained(
        mc.directory,
        device_map="auto",
        torch_dtype=torch.float16,
        revision="float16",
    )
else:
    model = AutoModelForCausalLM.from_pretrained(
        mc.directory,
        torch_dtype="auto",
        device_map="auto"
    )

print(model)
print(model.config)

def critique(row) -> str:
    """Inference method and response extraction based on https://huggingface.co/Qwen/Qwen2-7B-Instruct"""
    prompt = template.render(essay=str(row["full_text"])[:mc.char_limit])
    messages = [{"role": "user", "content": prompt}]
    text = tokenizer.apply_chat_template(messages, tokenize=False, add_generation_prompt=True)
    model_inputs = tokenizer(text, truncation=True, return_tensors="pt").to(device)
    model_outputs = model.generate(
        **model_inputs, 
        max_new_tokens=max_new_tokens,
        do_sample=False,
        #temperature=0.01,
        #top_p=0.95,
        #top_k=40,
        repetition_penalty=1.1,
        pad_token_id=tokenizer.eos_token_id,
    )
    generated_ids = [output_ids[len(input_ids):] for input_ids, output_ids in zip(model_inputs.input_ids, model_outputs)]
    res = tokenizer.batch_decode(generated_ids, skip_special_tokens=True)[0]
    res = basic(res)
    if len(res)==1 and res.isdigit():
        res = f"SCORE OF {res}"
    return res


df[critique_col] = df.progress_apply(critique, axis=1)

pdx.value_counts(df[critique_col]).head(30)

# Review Data

df.info()

cols = [critique_col, "score", "full_text"]
df[cols].head(10)

df.to_parquet(f"output/{input_filename}", index=False)
assert df.notna().all(axis=None)

tim.stop()
print(f"Total time taken {str(tim.elapsed)}")
