==> 30_combine_split.ipynb <==
import os
import json
import gc
from datetime import datetime
from pathlib import Path
import numpy as np
import pandas as pd
from sklearn.model_selection import StratifiedKFold, StratifiedGroupKFold
from tqdm import tqdm
from typing import Dict, List, Set, Tuple, NamedTuple, Callable
import scml
from scml import pandasx as pdx

version = "06"
files = [
    Path("input/comp_02.parquet"),
    Path("input/persuade_02.parquet"),
]
n_splits = 15  # 15 topics exist in corpus
persuade_topic_classification = False
if persuade_topic_classification:
    n_splits = 25

tim = scml.Timer()
tim.start()
os.environ["TOKENIZERS_PARALLELISM"] = "false"
percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()
scml.seed_everything()
info = np.iinfo(np.int16)
print(f"int16, min={info.min}, max={info.max}")

cols = [
    "essay_id", 
    "score", 
    #"ctq_Qwen2-1.5B-Instruct", 
    "ctq_3_Qwen2-1.5B-Instruct", 
    "topic", 
    "full_text"
]
cmb = None
for filepath in files:
    df = pd.read_parquet(filepath)
    df = df[cols]
    df["source"] = filepath.stem
    if cmb is None:
        cmb = df
    else:
        cmb = pd.concat([cmb, df], ignore_index=True)
df = cmb
df.info()

more = len(df)
df = df.drop_duplicates(["full_text"], keep="first", ignore_index=True)
print(f"{more - len(df):,} rows dropped: duplicate text")

pdx.value_counts(df["topic"])

# Train/Test Split

dummy = np.zeros(len(df))
if persuade_topic_classification:
    splitter = StratifiedKFold(n_splits=n_splits, shuffle=True)
    for ti, vi in splitter.split(dummy, y=df["topic"]):
        tra = df.iloc[ti]
        val = df.iloc[vi]
        break
else:
    splitter = StratifiedGroupKFold(n_splits=n_splits, shuffle=True)
    for ti, vi in splitter.split(dummy, y=df["score"], groups=df["topic"]):
        tra = df.iloc[ti]
        val = df.iloc[vi]
        break
tra.info()

val.info()

tra.to_parquet(f"output/tra_{version}.parquet", index=False)
val.to_parquet(f"output/val_{version}.parquet", index=False)
assert tra.notna().all(axis=None)
assert val.notna().all(axis=None)

tim.stop()
print(f"Total time taken {str(tim.elapsed)}")
