==> 40_features.ipynb <==
Features for Meta-Learner in Stacking Ensemble

from pathlib import Path
import os
import sys
import gc
import shutil
import json
import math
import uuid
from collections import defaultdict
import numpy as np
import pandas as pd
from transformers import AutoTokenizer, AutoModelForSequenceClassification
import torch
from tqdm import tqdm
import matplotlib.pyplot as plt
from typing import Dict, List, Tuple, NamedTuple, Callable, Iterable, Set, Optional, Any
import textstat
import scml
from scml import pandasx as pdx
import lalaes2 as mylib

version = "01"
text_col = "clean_text"
features = []

class ModelConf(NamedTuple):
    name: str
    directory: Path
    oof_pred_file: Path
    model_max_length: int
    batch_size: int
    model_class: str = "auto"


models_conf: List[ModelConf] = [
    ModelConf(
        name="deberta_base", 
        directory=Path("models/aes2/deberta_v3_base/20240615_063400"),
        oof_pred_file = Path("models/aes2/deberta_v3_base/20240616_110724/oof.npy"),
        model_max_length=512,
        batch_size=128,
    ),
]

tim = scml.Timer()
tim.start()
percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
os.environ["TOKENIZERS_PARALLELISM"] = "false"
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()
scml.seed_everything()

tra = pd.read_parquet(f"input/tra_{version}.parquet")
tra = tra.drop(columns=["source", "str_level"])
tra.info()

val = pd.read_parquet(f"input/val_{version}.parquet")
val = val.drop(columns=["source", "str_level"])
val.info()

# Base Model Inference

for mc in models_conf:
    print(mc.name)
    val[mc.name] = mylib.predict_holistic_score(
        ds=mylib.Aes2Dataset(
            tokenizer=AutoTokenizer.from_pretrained(mc.directory, model_max_length=mc.model_max_length),
            texts=val[text_col].tolist(),
        ),
        model=AutoModelForSequenceClassification.from_pretrained(mc.directory),
        batch_size=mc.batch_size,
        device=torch.device("cuda:1"),
        progress_bar=True,
    )
    with open(mc.oof_pred_file, "rb") as f:
        tra[mc.name] = np.load(f)
    features.append(mc.name)

# Character & Word-level Features

#%%time
#col = "cw_len"
#tra[col] = tra[text_col].str.len()
#val[col] = val[text_col].str.len()
#features.append(col)

def digit_frac(row) -> float:
    return mylib.digit_frac(row[text_col])


def letter_frac(row) -> float:
    return mylib.letter_frac(row[text_col])


def space_frac(row) -> float:
    return mylib.space_frac(row[text_col])


def punc_frac(row) -> float:
    return mylib.punc_frac(row[text_col])


def upper_frac(row) -> float:
    return mylib.upper_frac(row[text_col])


def repeat_char_frac(row) -> float:
    return mylib.repeat_char_frac(row[text_col])


def repeat_substring_frac(row) -> float:
    return mylib.repeat_substring_frac(row[text_col])


def unique_word_frac(row) -> float:
    return mylib.unique_word_frac(row[text_col])


sf = mylib.StopwordFraction()


def stopword_frac(row) -> float:
    return sf(row[text_col])



cw_fns: List[Tuple[str, Callable]] = [
    ("cw_digit_frac", digit_frac),
    ("cw_letter_frac", letter_frac),
    ("cw_space_frac", space_frac),
    ("cw_punc_frac", punc_frac),
    ("cw_upper_frac", upper_frac),
    ("cw_repeat_char_frac", repeat_char_frac),
    ("cw_repeat_substring_frac", repeat_substring_frac),
    ("cw_unique_word_frac", unique_word_frac),
    ("cw_stopword_frac", stopword_frac),
]   
#for col, fn in cw_fns:
#    print(col)
#    tra[col] = tra.progress_apply(fn, axis=1)
#    val[col] = val.progress_apply(fn, axis=1)
#    features.append(col)
#tra[features] = tra[features].astype(np.float32)
#val[features] = val[features].astype(np.float32)

# Textstat Features

def syllable_count(row) -> int:
    return textstat.syllable_count(row[text_col])


def lexicon_count(row) -> int:
    return textstat.lexicon_count(row[text_col])


def sentence_count(row) -> int:
    return textstat.sentence_count(row[text_col])


def syllables_per_word(row) -> float:
    return row["ts_syllable_count"] / (row["ts_lexicon_count"] + 1)


def syllables_per_sent(row) -> float:
    return row["ts_syllable_count"] / (row["ts_sentence_count"] + 1)


def words_per_sent(row) -> float:
    return row["ts_lexicon_count"] / (row["ts_sentence_count"] + 1)


def flesch_reading_ease(row) -> float:
    return textstat.flesch_reading_ease(row[text_col])


def flesch_kincaid_grade(row) -> float:
    return textstat.flesch_kincaid_grade(row[text_col])


def gunning_fog(row) -> float:
    return textstat.gunning_fog(row[text_col])


def smog_index(row) -> float:
    return textstat.smog_index(row[text_col])


def automated_readability_index(row) -> float:
    return textstat.automated_readability_index(row[text_col])


def coleman_liau_index(row) -> float:
    return textstat.coleman_liau_index(row[text_col])


def linsear_write_formula(row) -> float:
    return textstat.linsear_write_formula(row[text_col])


def dale_chall_readability_score(row) -> float:
    return textstat.dale_chall_readability_score(row[text_col])


def text_standard(row) -> float:
    return textstat.text_standard(row[text_col], float_output=True)


def mcalpine_eflaw(row) -> float:
    return textstat.mcalpine_eflaw(row[text_col])


textstat_fns: List[Tuple[str, Callable]] = [
    ("ts_syllable_count", syllable_count),
    ("ts_lexicon_count", lexicon_count),
    ("ts_sentence_count", sentence_count),
    ("ts_syllables_per_word", syllables_per_word),
    ("ts_syllables_per_sent", syllables_per_sent),
    ("ts_words_per_sent", words_per_sent),
    ("ts_flesch_reading_ease", flesch_reading_ease),
    ("ts_flesch_kincaid_grade", flesch_kincaid_grade),
    ("ts_gunning_fog", gunning_fog),
    ("ts_smog_index", smog_index),
    ("ts_automated_readability_index", automated_readability_index),
    ("ts_coleman_liau_index", coleman_liau_index),
    ("ts_linsear_write_formula", linsear_write_formula),
    ("ts_dale_chall_readability_score", dale_chall_readability_score),
    ("ts_text_standard", text_standard),
    ("ts_mcalpine_eflaw", mcalpine_eflaw),
]    
for col, fn in textstat_fns:
    print(col)
    tra[col] = tra.progress_apply(fn, axis=1)
    val[col] = val.progress_apply(fn, axis=1)
    if col.endswith("_count"):
        continue
    features.append(col)
tra[features] = tra[features].astype(np.float32)
val[features] = val[features].astype(np.float32)

# Review Data

tra[features] = tra[features].astype(np.float32)
val[features] = val[features].astype(np.float32)
tra = tra.drop(columns=["clean_text"])
val = val.drop(columns=["clean_text"])
tra.info()

val.info()

tra.describe(percentiles=percentiles)

val.describe(percentiles=percentiles)

features.sort()
with open(f"output/features_{version}.json", "w") as f:
    json.dump({"feature_names": features}, f)

tra.to_parquet(f"output/mtra_{version}.parquet", index=False)
val.to_parquet(f"output/mval_{version}.parquet", index=False)
assert tra.notna().all(axis=None)
assert val.notna().all(axis=None)

tim.stop()
print(f"Total time taken {str(tim.elapsed)}")
