==> 80_evaluation.ipynb <==
import os
import json
from pathlib import Path
import numpy as np
import pandas as pd
import torch
from tqdm import tqdm
from transformers import AutoTokenizer, AutoModelForSequenceClassification
import scml
from scml import pandasx as pdx
import lalaes2 as mylib

model_dir = Path("models/aes2/deberta_v3_base/20240628_111207")
validation_data_file = Path("input/val_06.parquet")
model_max_length = 768
batch_size = 32
critique_column = "ctq_3_Qwen2-1.5B-Instruct"
#model_class = "CustomDebertaV2ForTokenClassification"
model_class = "auto"
if model_max_length==768:
    batch_size = 16

tim = scml.Timer()
tim.start()
os.environ["TOKENIZERS_PARALLELISM"] = "false"
percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()
scml.seed_everything()
info = np.iinfo(np.int16)
print(f"int16, min={info.min}, max={info.max}")

device = torch.device("cpu")
if torch.backends.mps.is_available():
    device = torch.device("mps")
    batch_size = 128
    print("mps")
elif torch.cuda.is_available():
    device = torch.device("cuda:1")
    for i in range(torch.cuda.device_count()):
        print(f"device={i}, {torch.cuda.get_device_name(i)}")
        print('Mem Allocated:', round(torch.cuda.memory_allocated(i)/1024**3,1), 'GB')
        print('Mem Cached:   ', round(torch.cuda.memory_reserved(i)/1024**3,1), 'GB')
else:
    print("cpu")

df = pd.read_parquet(validation_data_file)
df.info()

tokenizer = AutoTokenizer.from_pretrained(model_dir, model_max_length=model_max_length)
print(tokenizer)

%%time
if model_class=="CustomDebertaV2ForTokenClassification":
    #model = CustomDebertaV2ForTokenClassification.from_pretrained(model_dir)
    raise ValueError
else:
    model = AutoModelForSequenceClassification.from_pretrained(model_dir)
print(model)

%%time
res = mylib.evaluation_aes2(
    ds=mylib.Aes2Dataset(
        tokenizer=tokenizer,
        critiques=df[critique_column].tolist(),
        texts=df["full_text"].tolist(),
        labels=df["score"].tolist(),
    ),
    model=model,
    batch_size=batch_size,
    device=device,
    progress_bar=True,
)

print(model_dir)
print(res["thresholds"])
print(json.dumps(res, indent=2))

tim.stop()
print(f"Total time taken {str(tim.elapsed)}")
