==> 91_llm_playground.ipynb <==
from pathlib import Path
import os
import sys
import gc
import re
import shutil
import json
import math
import jinja2
from collections import defaultdict
import numpy as np
import pandas as pd
import bitsandbytes
import accelerate
import torch
from transformers import AutoTokenizer, AutoModelForCausalLM, BitsAndBytesConfig
from tqdm import tqdm
import matplotlib.pyplot as plt
from typing import Dict, List, Tuple, NamedTuple, Callable, Iterable, Set, Optional, Any
import scml
print(f"accelerate={accelerate.__version__}, bitsandbytes={bitsandbytes.__version__}")

model_dir = Path("huggingface/mistralai/Mixtral-8x7B-Instruct-v0.1")
model_max_length = 4096
seed = 422
environment = jinja2.Environment()

tim = scml.Timer()
tim.start()
percentiles=[.01, .05, .1, .2, .3, .4, .5, .6, .7, .8, .9, .95, .99]
os.environ["TOKENIZERS_PARALLELISM"] = "false"
pd.set_option("max_info_columns", 9999)
pd.set_option("display.max_columns", 9999)
pd.set_option("display.max_rows", 9999)
pd.set_option('max_colwidth', 9999)
tqdm.pandas()
scml.seed_everything(seed)

device = torch.device("cpu")
if torch.cuda.is_available():
    device = torch.device("cuda")
    for i in range(torch.cuda.device_count()):
        print(f"device={i}, {torch.cuda.get_device_name(i)}")
        print('Mem Allocated:', round(torch.cuda.memory_allocated(i)/1024**3,1), 'GB')
        print('Mem Cached:   ', round(torch.cuda.memory_reserved(i)/1024**3,1), 'GB')
else:
    print("cpu")

df = pd.read_csv("input/persuade20/persuade_2.0_human_scores_demo_id_github.csv", low_memory=False)
df["full_text_len"] = df["full_text"].str.len()
df.info()

title_to_prompt = {}
title_to_exemplars = {}
for t in df.itertuples():
    prompt_name = str(getattr(t, "prompt_name"))
    assignment = str(getattr(t, "assignment"))
    title_to_prompt[prompt_name] = assignment
for title in title_to_prompt.keys():
    mask = (df["prompt_name"]==title) & (df["holistic_essay_score"]==6)
    tmp = df[mask].sort_values(["full_text_len"], ascending=[True])
    title_to_exemplars[title] = tmp["full_text"].tolist()
print(json.dumps(title_to_prompt, indent=2))

tokenizer = AutoTokenizer.from_pretrained(model_dir, model_max_length=model_max_length)
tokenizer.pad_token = tokenizer.eos_token
print(f"{repr(tokenizer)}\nmodel_input_names={tokenizer.model_input_names}")
print(f"pad_token_id={tokenizer.pad_token_id}")

%%time
quantization_config = BitsAndBytesConfig(
    load_in_4bit=True,
    bnb_4bit_quant_type="nf4",
    bnb_4bit_compute_dtype=torch.bfloat16,
    bnb_4bit_use_double_quant=True,
)
model = AutoModelForCausalLM.from_pretrained(
    str(model_dir),
    device_map = "auto",
    quantization_config=quantization_config,
)

print(model)

def generate(prompt: str) -> str:
    inputs = tokenizer(prompt, return_tensors="pt").to(device)
    outputs = model.generate(
        **inputs, 
        max_new_tokens=128,
        do_sample=True,
        temperature=1.0,
        top_p=0.95,
        top_k=100,
        repetition_penalty=1.1,
        pad_token_id=tokenizer.eos_token_id,
    )
    generated_text = tokenizer.batch_decode(outputs, skip_special_tokens=True)[0]
    generated_text = generated_text.split("[/INST]")[1]
    generated_text = generated_text.strip()
    return generated_text

template = environment.from_string(
"""[INST] You are a teacher setting an essay question for students in Grade 6-12.
Given the student's essay, output the essay question between opening tag [QUESTION] and closing tag [/QUESTION].
Output a short title of the essay question between opening tag [TITLE] and closing tag [/TITLE].
[ESSAY]{{ essay1 }}[/ESSAY]
[QUESTION]{{ question1 }}[/QUESTION]
[TITLE]{{ title1 }}[/TITLE]
[ESSAY]{{ essay2 }}[/ESSAY]
[QUESTION]{{ question2 }}[/QUESTION]
[TITLE]{{ title2 }}[/TITLE]
[ESSAY]{{ essay0 }}[/ESSAY]
[QUESTION][/QUESTION]
[TITLE][/TITLE][/INST]"""
)

%%time
prompt = template.render(
    essay0=title_to_exemplars["Cell phones at school"][0],
    essay1=title_to_exemplars["Exploring Venus"][0],
    question1=title_to_prompt["Exploring Venus"],
    title1="Exploring Venus",
    essay2=title_to_exemplars["Does the electoral college work?"][0],
    question2=title_to_prompt["Does the electoral college work?"],
    title2="Does the electoral college work?",
)
output = generate(prompt)
question = ""
title = ""
print(f"{prompt}\n\n=====  END OF PROMPT  =====\n\n{output}")

pattern = re.compile(r"\[QUESTION\](.+)\[/QUESTION\]", re.IGNORECASE)
for m in pattern.finditer(output):
    if m[0] == "":
        continue
    question = m[1]
    break
pattern = re.compile(r"\[TITLE\](.+)\[/TITLE\]", re.IGNORECASE)
for m in pattern.finditer(output):
    if m[0] == "":
        continue
    title = m[1]
    break
print(f"title={title}\nquestion={question}")

tim.stop()
print(f"Total time taken {str(tim.elapsed)}")
