==> 93_hf_hub_download.ipynb <==
from huggingface_hub import snapshot_download

#repo_id = "google/gemma-1.1-2b-it"
#repo_id = "meta-llama/Meta-Llama-3-8B"
repo_id = "Qwen/Qwen2-7B-Instruct"
snapshot_download(
    repo_id=repo_id, 
    allow_patterns=["*.model", "*.json", "*.safetensors"],
    local_dir=f"/mnt/c/huggingface/{repo_id}",
)
